+++ frontend/src/components/NavBar.js
@@ -1,5 +1,5 @@
 import { NavLink } from "react-router-dom"
-import { useState, useEffect } from "react"
+import { useState } from "react"
 
 export const NavBar = () => {
